[ContentFragment] Improve the Content Fragment feature #120
