+++ extension/contentfragment/content/pom.xml
@@ -23,17 +23,17 @@
     <parent>
         <groupId>com.adobe.cq</groupId>
         <artifactId>core.wcm.components.sandbox.parent</artifactId>
-        <version>0.0.1-SNAPSHOT</version>
+        <version>2.0.0</version>
         <relativePath>../../../parent/pom.xml</relativePath>
     </parent>
 
     <!-- ====================================================================== -->
     <!-- P R O J E C T  D E S C R I P T I O N                                   -->
     <!-- ====================================================================== -->
-    <artifactId>core.wcm.components.sandbox.extension.contentfragment.content</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <artifactId>core.wcm.components.extension.contentfragment.content</artifactId>
+    <version>0.0.9-SNAPSHOT</version>
     <packaging>content-package</packaging>
-    <name>Adobe Experience Manager Core WCM Components Extension - Structured Content Fragment Content - Sandbox</name>
+    <name>Adobe Experience Manager Core WCM Components Extension - Structured Content Fragment Content</name>
 
     <scm>
         <connection>scm:git:https://git.corp.adobe.com/CQ/aem-core-wcm-components.git</connection>
@@ -151,8 +151,8 @@
     <dependencies>
         <dependency>
             <groupId>com.adobe.cq</groupId>
-            <artifactId>core.wcm.components.sandbox.extension.contentfragment.bundle</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
+            <artifactId>core.wcm.components.extension.contentfragment.bundle</artifactId>
+            <version>0.0.6</version>
             <scope>provided</scope>
         </dependency>
