+++ extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/components/contentfragment/v1/contentfragment/.content.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
+    cq:icon="documentFragment"
+    jcr:description="Displays the elements of a Content Fragment or a selection thereof"
+    jcr:primaryType="cq:Component"
+    jcr:title="Content Fragment (v1)"
+    componentGroup=".core-wcm"/>
