--- extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/sandbox/components/contentfragment/v1/contentfragment/.content.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:Component"
-    jcr:title="Content Fragment (v1 - sandbox)"
-    jcr:description="Displays the elements of a Content Fragment or a selection thereof"
-    componentGroup=".core-wcm"/>
