--- extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/sandbox/components/contentfragment/v1/contentfragment/README.md
@@ -1,50 +1,0 @@
-<!--
-Copyright 2017 Adobe Systems Incorporated
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-Content Fragment (v1 - sandbox extension)
-====
-Content Fragment component written in HTL that displays the elements of a Content Fragment or a selection thereof.
-
-## Features
-* Displays the elements of a Content Fragment as an HTML description list
-* By default renders all elements of a Content Fragment
-* Can be configured to render a subset of the elements in a specific order
-
-### Use Object
-The Content Fragment component uses the `com.adobe.cq.wcm.core.components.sandbox.extension.contentfragment.models.ContentFragment` Sling model as its Use-object.
-
-### Edit dialog properties
-The following JCR properties are used:
-
-1. `./fragmentPath` - defines the path to the Content Fragment to be rendered
-2. `./elementNames` - multi-valued property defining the elements to be rendered and in which order (optional: if not present, all elements are rendered)
-3. `./variationName` - defines the variation to use to render the elements (optional: if not present, the master variation is used)
-
-## BEM description
-```
-BLOCK cmp-contentfragment
-    ELEMENT cmp-contentfragment__element
-        MOD cmp-contentfragment__element--<name>
-    ELEMENT cmp-contentfragment__element-label
-    ELEMENT cmp-contentfragment__element-value
-```
-
-## Information
-* **Vendor**: Adobe
-* **Version**: v1 - sandbox extension
-* **Compatibility**: AEM 6.3
-* **Status**: production-ready
-* **Documentation**: [https://www.adobe.com/go/aem\_cmp\_contentfragment\_v1](https://www.adobe.com/go/aem_cmp_contentfragment_v1)
