--- extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/sandbox/components/contentfragment/v1/contentfragment/_cq_dialog/.content.xml
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:granite="http://www.adobe.com/jcr/granite/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0"
-    jcr:primaryType="nt:unstructured"
-    jcr:title="Content Fragment"
-    sling:resourceType="cq/gui/components/authoring/dialog"
-    extraClientlibs="[core.wcm.extension.sandbox.components.contentfragment.v1.dialog]"
-    helpPath="https://www.adobe.com/go/aem_cmp_contentfragment_v1">
-    <content
-        jcr:primaryType="nt:unstructured"
-        sling:resourceType="granite/ui/components/coral/foundation/container"
-        granite:class="cmp-contentfragment__edit-dialog">
-        <items jcr:primaryType="nt:unstructured">
-            <tabs
-                jcr:primaryType="nt:unstructured"
-                sling:resourceType="granite/ui/components/coral/foundation/tabs"
-                maximized="{Boolean}true">
-                <items jcr:primaryType="nt:unstructured">
-                    <properties
-                        jcr:primaryType="nt:unstructured"
-                        jcr:title="Properties"
-                        sling:resourceType="granite/ui/components/coral/foundation/container"
-                        margin="{Boolean}true">
-                        <items jcr:primaryType="nt:unstructured">
-                            <column
-                                jcr:primaryType="nt:unstructured"
-                                sling:resourceType="granite/ui/components/coral/foundation/container">
-                                <items jcr:primaryType="nt:unstructured">
-                                    <fragmentPath
-                                        jcr:primaryType="nt:unstructured"
-                                        sling:resourceType="dam/cfm/components/cfpicker"
-                                        name="./fragmentPath"
-                                        fieldLabel="Content Fragment"
-                                        emptyText="Enter or select Content Fragment"
-                                        pickerTitle="Select Content Fragment"
-                                        rootPath="/content/dam"/>
-                                    <variationName
-                                        jcr:primaryType="nt:unstructured"
-                                        sling:resourceType="granite/ui/components/coral/foundation/form/select"
-                                        name="./variationName"
-                                        fieldLabel="Variation"
-                                        fieldDescription="Specify which variation to display">
-                                        <datasource
-                                            jcr:primaryType="nt:unstructured"
-                                            sling:resourceType="core/wcm/extension/sandbox/components/contentfragment/v1/datasource/variations"
-                                            componentPath="${requestPathInfo.suffix}"
-                                            fragmentPath="${param.fragmentPath}"/>
-                                        <granite:data
-                                            jcr:primaryType="nt:unstructured"
-                                            field-path="${requestPathInfo.resourcePath}"/>
-                                    </variationName>
-                                    <elementNames
-                                        jcr:primaryType="nt:unstructured"
-                                        sling:resourceType="granite/ui/components/coral/foundation/form/multifield"
-                                        fieldLabel="Elements"
-                                        fieldDescription="Specify which elements to display or leave empty to display all elements.">
-                                        <field
-                                            jcr:primaryType="nt:unstructured"
-                                            sling:resourceType="granite/ui/components/coral/foundation/form/select"
-                                            emptyText="Select"
-                                            fieldLabel="Element"
-                                            name="./elementNames">
-                                            <datasource
-                                                jcr:primaryType="nt:unstructured"
-                                                sling:resourceType="core/wcm/extension/sandbox/components/contentfragment/v1/datasource/elements"
-                                                componentPath="${requestPathInfo.suffix}"
-                                                fragmentPath="${param.fragmentPath}"/>
-                                        </field>
-                                        <granite:data
-                                            jcr:primaryType="nt:unstructured"
-                                            field-path="${requestPathInfo.resourcePath}"/>
-                                    </elementNames>
-                                </items>
-                            </column>
-                        </items>
-                    </properties>
-                </items>
-            </tabs>
-        </items>
-    </content>
-</jcr:root>
