--- extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/sandbox/components/contentfragment/v1/contentfragment/_cq_editConfig.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:EditConfig">
-    <cq:inplaceEditing
-        jcr:primaryType="cq:InplaceEditingConfig"
-        active="{Boolean}true"
-        editorType="contentfragment"/>
-    <cq:dropTargets jcr:primaryType="nt:unstructured">
-        <contentfragment
-            jcr:primaryType="cq:DropTargetConfig"
-            accept="[application/vnd.adobe.contentfragment]"
-            groups="[media]"
-            propertyName="./fragmentPath">
-            <parameters
-                jcr:primaryType="nt:unstructured"
-                elementNames=""
-                variationName=""/>
-        </contentfragment>
-    </cq:dropTargets>
-    <cq:listeners
-        jcr:primaryType="cq:EditListenersConfig"
-        afterinsert="function (properties) { $(document).trigger('cq-contentfragment-insert'); }"
-        afteredit="function (properties) { $(document).trigger('cq-contentfragment-edit'); }"
-        afterdelete="function (properties) { $(document).trigger('cq-contentfragment-delete'); }"/>
-</jcr:root>
