--- extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/sandbox/components/contentfragment/v1/contentfragment/clientlibs/editor/.content.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="sling:Folder"/>
