--- extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/sandbox/components/contentfragment/v1/contentfragment/clientlibs/editor/authoring/.content.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:ClientLibraryFolder"
-    categories="[core.wcm.extension.sandbox.components.contentfragment.v1.authoring,cq.authoring.editor.hook]"
-    dependencies="[cq.authoring.editor.core]"/>
