--- extension/contentfragment/content/src/content/jcr_root/apps/core/wcm/extension/sandbox/components/contentfragment/v1/contentfragment/contentfragment.html
@@ -1,30 +1,0 @@
-<!--/*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  ~ Copyright 2017 Adobe Systems Incorporated
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
-<!--/* need this outer test for the fragment path, as it is a required property for instantiating the model */-->
-<sly data-sly-test="${properties.fragmentPath}">
-    <dl data-sly-use.fragment="com.adobe.cq.wcm.core.components.sandbox.extension.contentfragment.models.ContentFragment"
-        data-sly-test="${!fragment.elements.isEmpty}"
-        data-path="${properties.fragmentPath}"
-        data-type="${fragment.type}"
-        data-json="${!wcmmode.disabled && fragment.editorJSON}"
-        class="cmp-contentfragment${!wcmmode.disabled ? ' cq-dd-contentfragment' : ''}"
-        data-sly-use.elementTemplate="element.html"
-        data-sly-list.element="${fragment.elements}">
-        <sly data-sly-call="${elementTemplate.element @ element=element}"></sly>
-    </dl>
-</sly>
-<sly data-sly-use.template="core/wcm/components/commons/v1/templates.html"
-     data-sly-call="${template.placeholder @ isEmpty=!properties.fragmentPath || fragment.elements.isEmpty, classAppend='cq-dd-contentfragment'}"></sly>
