+++ extension/pom.xml
@@ -21,17 +21,17 @@
     <!-- P A R E N T  P R O J E C T  D E S C R I P T I O N                      -->
     <!-- ====================================================================== -->
     <parent>
-        <artifactId>core.wcm.components.sandbox.parent</artifactId>
+        <artifactId>core.wcm.components.parent</artifactId>
         <groupId>com.adobe.cq</groupId>
-        <version>0.0.1-SNAPSHOT</version>
+        <version>1.9.28</version>
         <relativePath>../parent/pom.xml</relativePath>
     </parent>
 
     <!-- ====================================================================== -->
     <!-- P R O J E C T  D E S C R I P T I O N                                   -->
     <!-- ====================================================================== -->
-    <artifactId>core.wcm.components.sandbox.extension</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <artifactId>core.wcm.components.extension</artifactId>
+    <version>0.0.3-SNAPSHOT</version>
     <packaging>content-package</packaging>
 
     <name>Adobe Experience Manager Core WCM Components Extension Package</name>
@@ -64,14 +64,14 @@
                     <embeddeds>
                         <embedded>
                             <groupId>${project.groupId}</groupId>
-                            <artifactId>core.wcm.components.sandbox.extension.contentfragment.bundle</artifactId>
+                            <artifactId>core.wcm.components.extension.contentfragment.bundle</artifactId>
                             <filter>true</filter>
                         </embedded>
                     </embeddeds>
                     <subPackages>
                         <subPackage>
                             <groupId>${project.groupId}</groupId>
-                            <artifactId>core.wcm.components.sandbox.extension.contentfragment.content</artifactId>
+                            <artifactId>core.wcm.components.extension.contentfragment.content</artifactId>
                             <filter>true</filter>
                         </subPackage>
                     </subPackages>
@@ -156,13 +156,13 @@
     <dependencies>
         <dependency>
             <groupId>com.adobe.cq</groupId>
-            <artifactId>core.wcm.components.sandbox.extension.contentfragment.bundle</artifactId>
-            <version>0.0.3-SNAPSHOT</version>
+            <artifactId>core.wcm.components.extension.contentfragment.bundle</artifactId>
+            <version>0.0.6</version>
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>core.wcm.components.sandbox.extension.contentfragment.content</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
+            <artifactId>core.wcm.components.extension.contentfragment.content</artifactId>
+            <version>0.0.8</version>
             <type>zip</type>
             <scope>provided</scope>
         </dependency>
